Port netatalk-classic patches between Feb 2022 and Jan 2023 (#23)
* Use correct BSD make syntax.

* Use fseek() here too.

* Use explicitly sized types.

* Ignore another working directory.

* Ignore all .AppleDouble directories.

* Properly handle possible use-after-free.

* Cast to proper type.  Fixes compilation on 32-bit hosts.

* Use long long in print string, to placate 32-bit compilation.

* Include stdint.h for the uintptr_t data type

---------

+++ etc/afpd/uam.c
@@ -11,6 +11,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 /* STDC check */
 #include <string.h>
@@ -340,8 +341,8 @@ int uam_random_string(AFPObj * obj, char *buf, int len)
 
 		if (gettimeofday(&tv, &tz) < 0)
 			return -1;
-		srandom(tv.tv_sec + (unsigned long) obj +
-			(unsigned long) obj->handle);
+		srandom(tv.tv_sec + (uintptr_t) obj +
+			(uintptr_t) obj->handle);
 		for (i = 0; i < len; i += sizeof(result)) {
 			result = random();
 			memcpy(buf + i, &result, sizeof(result));
